requirements.txt
strands-agents 
strands-agents[openai] 
strands-agents[litellm] 
strands-agents-tools 
strands-agents-builder 
a2a-sdk 
a2a-sdk[sqlite] 
strands-nvidia-nim 
fastmcp 
python-dotenv 
rich
boto3
langchain
langchain-community
langchain-aws
faiss-cpu
gitpython
requests
markdown
reportlab
pypdf2